fix: uninstall dotenv and refactor tests
