+++ .env
@@ -1,4 +1,1 @@
 BABEL_ENV=production
-COASTER_BASE_URL=https://api.coaster.cloud
-COASTER_ACL=false
-COASTER_DEFAULT_LANGUAGE=en
