+++ package.json
@@ -19,7 +19,6 @@
   "homepage": "https://github.com/migo315/coaster.js#readme.md",
   "dependencies": {
     "axios": "^0.18.0",
-    "dotenv": "^8.1.0",
     "lodash": "^4.17.11"
   },
   "devDependencies": {
