+++ src/client.js
@@ -9,18 +9,17 @@
 
 // Imports
 const Axios = require('axios');
-const _ = require('lodash');
-const { baseUrl, withAcl, language } = require('../config');
+const { baseUrl, withAcl } = require('../config');
+const { buildQuery } = require('./build-query');
 
 module.exports = {
-    withAcl,
 
     // Get parks list by optional config or null on failure
     async getParks(config = {}) {
         let result = null;
 
         try {
-            let response = await Axios.get(baseUrl + '/parks' + this.buildQuery(config));
+            let response = await Axios.get(baseUrl + '/parks' + buildQuery(config, withAcl));
             result = response.data;
         } catch (error) {
             console.error(error)
@@ -34,7 +33,7 @@ module.exports = {
         let result = null;
 
         try {
-            let response = await Axios.get(baseUrl + '/parks/' + uuid + this.buildQuery(config));
+            let response = await Axios.get(baseUrl + '/parks/' + uuid + buildQuery(config, withAcl));
             result = response.data;
         } catch (error) {
             console.error(error)
@@ -49,7 +48,7 @@ module.exports = {
 
         try {
             let response = await Axios.get(
-                baseUrl + '/parks/' + uuid  + '/waiting-times' + this.buildQuery(config)
+                baseUrl + '/parks/' + uuid  + '/waiting-times' + buildQuery(config, withAcl)
             );
             result = response.data;
         } catch (error) {
@@ -71,7 +70,7 @@ module.exports = {
         }
 
         try {
-            let response = await Axios.get(url + this.buildQuery(config));
+            let response = await Axios.get(url + buildQuery(config, withAcl));
             result = response.data;
         } catch (error) {
             console.error(error)
@@ -85,7 +84,7 @@ module.exports = {
         let result = null;
 
         try {
-            let response = await Axios.get(baseUrl + '/attractions/' + uuid + this.buildQuery(config));
+            let response = await Axios.get(baseUrl + '/attractions/' + uuid + buildQuery(config, withAcl));
             result = response.data;
         } catch (error) {
             console.error(error)
@@ -112,71 +111,4 @@ module.exports = {
 
         return result;
     },
-
-    // Build query
-    buildQuery(config) {
-        let query = [];
-
-        if (config.hasOwnProperty('page') && config.page) {
-            query.push('page=' + encodeURIComponent(config.page));
-        }
-
-        if (config.hasOwnProperty('itemsPerPage') && config.itemsPerPage) {
-            query.push('itemsPerPage=' + encodeURIComponent(config.itemsPerPage));
-        }
-
-        if (config.hasOwnProperty('searchTerm') && config.searchTerm) {
-            query.push('search=' + encodeURIComponent(config.searchTerm));
-        }
-
-        if (config.hasOwnProperty('language') && config.language) {
-            query.push('language=' + encodeURIComponent(config.language));
-        } else {
-            query.push('language=' + encodeURIComponent(language));
-        }
-
-        if (config.hasOwnProperty('sort') && config.sort) {
-            query.push('sort=' + encodeURIComponent(config.sort));
-        }
-
-        if (config.hasOwnProperty('regulation')) {
-            let regulation = _.merge(
-                {
-                    size: null,
-                    age: null
-                },
-                config.regulation
-            );
-
-            if (regulation.size !== null || regulation.age !== null) {
-                let size = (/^\d+$/.test(regulation.size)) ? regulation.size : 'null';
-                let age = (/^\d+$/.test(regulation.age)) ? regulation.age : 'null';
-
-                query.push('filter[]=' + encodeURIComponent(`regulation(${size},${age})`));
-            }
-        }
-
-        if (config.hasOwnProperty('facets')) {
-            config.facets.forEach(function (facet) {
-                query.push('facet[]=' + encodeURIComponent(facet));
-            });
-        }
-
-        if (config.hasOwnProperty('filters')) {
-            config.filters.forEach(function (filter) {
-                query.push('filter[]=' + encodeURIComponent(filter));
-            });
-        }
-
-        if (config.hasOwnProperty('founded') && config.founded !== null && (/^\d{4}$/.test(config.founded))) {
-            let year = config.founded;
-            query.push('filter[]=' + encodeURIComponent(`founded.year:${year}`));
-        }
-
-        if (this.withAcl === true) {
-            query.push('acl=true');
-        }
-
-        return query.length > 0 ? '?' + query.join('&') : '';
-    }
 };
